refactor: separate auth functions
Move auth repository functions into a separate file.
