+++ plural_app/lib/src/features/authentication/data/forms.dart
@@ -12,6 +12,7 @@ import 'package:plural_app/src/constants/fields.dart';
 import 'package:plural_app/src/constants/routes.dart';
 
 // Auth
+import "package:plural_app/src/features/authentication/data/auth_api.dart";
 import "package:plural_app/src/features/authentication/data/auth_repository.dart";
 import "package:plural_app/src/features/authentication/domain/constants.dart";
